centralize to one QueryBuilder
