+++ QueryBuilder.Test/QueryBuilder.Typed/Join.UnitTests.cs
@@ -6,7 +6,6 @@ namespace QueryBuilder.UnitTests.QueryBuilder.Typed
     using System;
     using global::Azure.DigitalTwins.Core;
     using Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder;
-    using Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder.Typed;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
 
     [TestClass]
