+++ QueryBuilder.Test/QueryBuilder.Typed/Select.UnitTests.cs
@@ -7,7 +7,6 @@ namespace QueryBuilder.UnitTests.QueryBuilder.Typed
     using System.Collections.Generic;
     using System.Linq;
     using Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder;
-    using Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder.Typed;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
 
     [TestClass]
