+++ QueryBuilder.Test/QueryBuilder.Typed/Where.UnitTests.cs
@@ -5,7 +5,6 @@ namespace QueryBuilder.UnitTests.QueryBuilder.Typed
 {
     using System;
     using Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder;
-    using Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder.Typed;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
 
     [TestClass]
