+++ QueryBuilder/Dynamic/Statements/CompoundWhereStatement.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder.Common.Statements
+namespace Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder.Dynamic.Statements
 {
     using System;
     using System.Collections.Generic;
