+++ QueryBuilder/Dynamic/Statements/JoinRelatedByStatement.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder.Common.Statements
+namespace Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder.Dynamic.Statements
 {
     using Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder.Common.Clauses;
