+++ QueryBuilder/Dynamic/Statements/TwinsWhereStatement.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder.Common.Statements
+namespace Microsoft.DigitalWorkplace.DigitalTwins.QueryBuilder.Dynamic.Statements
 {
     using System.Collections.Generic;
     using System.Linq;
